[Stroustrup_principles_and_practice] Added CH14P517EXERC14_FLTK example
CH14P517EXERC14_FLTK      / Add an operation to Binary_tree that adds text to a node. You may have to modify the design of Binary_tree to implement this elegantly. Choose a way to identify a node; for example, you might give a string "lrrlr" for navigating left, right, right, left, and right down a binary tree (the root node would match both an initial l and an initial r).