[Stroustrup_principles_and_practice][In_progress] Added CH15P548EXERC8_FLTK example
CH15P548EXERC8_FLTK       / Here is a collection of heights in centimeters together with the number of people in a group of that height (rounded to the nearest 5cm): (170,7), (175,9), (180,23), (185,17), (190,6), (195,1). How would you graph that data? If you can’t think of anything better, do a bar graph. Remember to provide axes and labels. Place the data in a file and read it from that file.