[Stroustrup_principles_and_practice] Added CH16P579EXERC4_FLTK example
CH16P579EXERC4_FLTK       / Make a menu with items that make a circle, a square, an equilateral triangle, and a hexagon, respectively. Make an input box (or two) for giving a coordinate pair, and place the shape made by pressing a menu item at that coordinate. Sorry, no drag and drop.